indic_trans_2/requirements.txt
transformers
IndicTransToolkit
scipy
sentencepiece
python-dotenv
huggingface_hub[hf_xet]
fastapi
uvicorn
torch --index-url https://download.pytorch.org/whl/nightly/cu130
torchvision --index-url https://download.pytorch.org/whl/nightly/cu130
